[Gold V] Title: 가로등, Time: 324 ms, Memory: 35504 KB, Score: 25 point -BaekjoonHub

+++ 백준/Gold/32069. 가로등/가로등.cc
@@ -16,8 +16,8 @@ int main() {
 		q.push({x,0});
 	}
 	while (k!= K) {
-		int now = q.front().first;
-		int cost = q.front().second;
+		long long now = q.front().first;
+		long long cost = q.front().second;
 		q.pop();
 		cout << cost << '\n';
 		k++;
